+++ changelog.md
@@ -10,6 +10,7 @@ Codion Change Log
 - DefaultInputPanelBuilder now uses the label from AbstractComponentBuilder if specified.
 - ColumnConditionPanel.createComponents() bug fixed, locked state validation was disabled a bit too late.
 - FilterTableCellRenderer.Configurer renamed Customizer.
+- ExceptionDialogBuilder, RuntimeException no longer a wrapping exception by default.
 ### is.codion.framework.domain
 - AttributeValidator added along with ValueAttributeDefinition.Builder.validator().
 - EntitySerializer bug fixed, key deserialization now fails if attribute is not found.